+++ src/engine.js
@@ -16,27 +16,44 @@ class Engine {
 
   constructor(props) {
     this.props = Object.assign({}, Engine.defaultProps, props);
-    this.tickId = null;
   }
 
   start() {
-    window.onmousemove = event => this._handleMouseMove(event);
-    this.state.started = true;
+    window.onmousemove = event => this._handleMovement(event);
+    window.ontouchmove = event => this._handleMovement(event);
+    this._setState({ started: true });
   }
 
   stop() {
-    this.state.started = false;
-    this.state.isMoving = false;
+    this._setState({ started: false, isMoving: false });
     window.onmousemove = noop;
+    window.ontouchmove = noop;
   }
 
-  _handleMouseMove(event) {
-    this.state.isMoving = true;
-    this.state.mouseX = event.x;
-    this.state.mouseY = event.y;
+  _setState(newState) {
+    this.state = Object.assign(this.state, newState);
+  }
+
+  _handleMovement(event) {
+    this._setState({ isMoving: true });
+
+    let mouseX, mouseY;
+    if (event.touches) {
+      mouseX = event.touches[0].clientX;
+      mouseY = event.touches[0].clientY;
+    } else {
+      mouseX = event.x;
+      mouseY = event.y;
+    }
+
+    this._setState({ mouseX, mouseY });
     this.props.onMouseChange(this.state);
 
-    this.state.isMoving = false;
+    this._setState({ isMoving: false });
+    this._handleMovementStopped();
+  }
+
+  _handleMovementStopped() {
     if (this.onMouseStopTimeoutId) clearTimeout(this.onMouseStopTimeoutId);
     this.onMouseStopTimeoutId = setTimeout(() => {
       if (!this.state.isMoving) this.props.onMouseStop();