+++ CHANGELOG.md
@@ -4,9 +4,14 @@
 
 ### :rocket: Enhancements
 
-* Increase `nav-links` font size ([@erremauro])
+* [#1] Support mobile devices by listening to `ontouchmove` events.
+* [#2] Increase `nav-links` font size.
 
-[@erremauro]: https://github.com/erremauro
+### :house: Internal
+
+* Use `nav-links` and `face` HTML element's bounding rect has references to
+calculate the area of interaction that defines when the eyes should move.
+* Update `Engine` to use `_setState` for updating state variables.
 
 [1.1.0]: https://github.com/erremauro/errews-homepage/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/erremauro/errews-homepage/tree/1.0.0